+++ generic/objects.js
@@ -1,5 +1,4 @@
 const { 
-    ENUM_STAT_NAMES,
     ENUM_EXPLORE_STATUS,
     ENUM_DUNGEON_DOOR_STATUS,
     ENUM_QUEST_STATUS,
